requirements.txt
openai==0.27.8
python-decouple==3.8
